Create Landmark Detection for Geotagging
Using the convolutional layers from a pretrained model and retraining the fully-connected layers over 100 epochs on images of 50 landmarks such as the Eiffel Tower, Trevi Fountain, and the Sydney Opera. The resulting neural network was used in a function to accurately return a geotag of the location the picture was taken in. It was done on the AWS SageMaker Notebook