Created a Neural Network using only Numpy
The neural network was used to predict how many bikes a bike-sharing company should have available depending on parameters such as the day of the week, the hour of the day, the temperature, the wind speed, humidity, etc. The hyperparameters at the bottom were tweaked until the training loss was able to reach 0.09 and the validation loss was 0.18.